+++ .storybook/css.story.js
@@ -2,6 +2,7 @@ import React from 'react';
 import _ from 'lodash';
 import { storiesOf } from '@storybook/react';
 import css from '/Users/drau/Sites/csswhat/css/oui.css.json';
+import css from '../data/csswhat/oui.css.json';
 
 const searchSelectors = function(e) {
   console.log(e);
@@ -70,16 +71,16 @@ stories
       </div>
 
     </div>
-  )})
-  .add('Search', () => {
-    console.log(css.selectors);
-    return (
-      <div>
-        <h2>Search</h2>
-        <input type="text" onKeyUp={ searchSelectors } />
-        <ul>
-          ...
-        </ul>
-      </div>
-    );
-  });
+)})
+.add('Search', () => {
+  console.log(css.selectors);
+  return (
+    <div>
+      <h2>Search</h2>
+      <input type="text" onKeyUp={ searchSelectors } />
+      <ul>
+        ...
+      </ul>
+    </div>
+  );
+});