+++ src/components/BlockList/BlockList.story.js
@@ -3,7 +3,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, text, boolean, select } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 
 import BlockList from './index.js';
 import Button from '../Button';
@@ -13,7 +12,6 @@ import OverlayWrapper from '../OverlayWrapper';
 const stories = storiesOf('BlockList', module);
 stories
   .addDecorator(withKnobs)
-  .addDecorator(withInfo)
   .addDecorator(story => (
     <div id="root-preview">
       {story()}