+++ src/components/Button/Button.story.js
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, select, text } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 import { action } from '@storybook/addon-actions';
 
 import ButtonRow from '../ButtonRow';
@@ -13,7 +12,6 @@ import ArrowsInline from '../ArrowsInline';
 const stories = storiesOf('Button', module);
 stories
   .addDecorator(withKnobs)
-  .addDecorator(withInfo)
   .addDecorator(story => (
     <div id="root-preview">
       {story()}