+++ src/components/Pagination/Pagination.story.js
@@ -3,14 +3,12 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, number } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 
 import Pagination from './index.js';
 
 const stories = storiesOf('Pagination', module);
 stories
   .addDecorator(withKnobs)
-  .addDecorator(withInfo)
   .addDecorator(story => (
     <div id="root-preview">
       {story()}