+++ src/components/Sortable/DragAndDrop.story.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { action } from '@storybook/addon-actions';
 import { DragDropContext } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
-import { withInfo } from '@storybook/addon-info';
 import { withKnobs, boolean } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 
@@ -56,7 +55,6 @@ const DNDContextContainer = DragDropContext(HTML5Backend)(({ children }) => (
 const stories = storiesOf('Sortable', module);
 stories
   .addDecorator(withKnobs)
-  .addDecorator(withInfo)
   .addDecorator(story => (
     <DNDContextContainer>
       <div id="root-preview">