+++ src/components/Spinner/Spinner.story.js
@@ -2,14 +2,12 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 
 import Spinner from './index.js';
 
 const stories = storiesOf('Spinner', module);
 stories
   .addDecorator(withKnobs)
-  .addDecorator(withInfo)
   .addDecorator(story => (
     <div id="root-preview">
       {story()}